feat: Input field is now renderless
