+++ packages/fe/components/form/fields/input.vue
@@ -1,64 +1,21 @@
 <template>
-  <div :class="['field field-input', state, { focused, empty, predictive, 'dropdown-open': dropdownOpen }]">
-
-    <div v-if="disabled" :class="['input', { disabled }]">
-      {{ value }}
-    </div>
-
-    <div v-else class="input-container">
-      <input
-        :id="fieldKey"
-        v-click-outside="closeDropdown"
-        :type="inputType"
-        :name="fieldKey"
-        :placeholder="placeholder"
-        :value="value"
-        :min="min"
-        :max="max"
-        :autocomplete="autocomplete"
-        :class="['input', state]"
-        @focus="toggleFocused(true)"
-        @blur="toggleFocused(false)"
-        @input="$emit('updateValue', $event.target.value)" />
-      <div v-if="typeof chars === 'number'" class="char-validation">
-        {{ chars }}
-      </div>
-    </div>
-
-    <div
-      v-if="predictive && predictiveOptions"
-      :class="['dropdown', { open: dropdownOpen }]">
-      <div v-if="predictiveFiltered" class="options-container">
-        <button
-          v-for="(option, index) in predictiveFiltered"
-          :key="index"
-          :class="['option', { selected: selected(option) }]"
-          @click="selectOption(option)">
-          {{ option[predictiveDisplayKey] }}
-        </button>
-      </div>
-      <div v-else class="none-found-placeholder">
-        No items found
-      </div>
-    </div>
-
-  </div>
+  <Input
+    :field="field"
+    :force-disabled="forceDisabled"
+    v-on="$listeners" />
 </template>
 
 <script>
-// =================================================================== Functions
-const preValidate = (instance, value, pre) => {
-  if (typeof pre !== 'string') { return }
-  const regex = new RegExp(pre)
-  if (regex.test(value)) { // value contains restricted characters
-    const stripped = value.replace(regex, '')
-    instance.$emit('updateValue', stripped)
-  }
-}
+// ===================================================================== Imports
+import Input from '@/modules/form/fields/input'
 
 // ====================================================================== Export
 export default {
   name: 'FieldInput',
+
+  components: {
+    Input
+  },
 
   props: {
     field: {
@@ -70,120 +27,6 @@ export default {
       required: false,
       default: false
     }
-  },
-
-  data () {
-    return {
-      focused: false,
-      dropdownOpen: false
-    }
-  },
-
-  computed: {
-    scaffold () {
-      return this.field.scaffold
-    },
-    inputType () {
-      return this.scaffold.inputType || 'text'
-    },
-    fieldKey () {
-      return this.field.fieldKey
-    },
-    label () {
-      return this.scaffold.label
-    },
-    placeholder () {
-      return this.scaffold.placeholder || 'Enter a value...'
-    },
-    autocomplete () {
-      return this.scaffold.autocomplete
-    },
-    required () {
-      return this.scaffold.required
-    },
-    disabled () {
-      return this.forceDisabled || this.scaffold.disabled
-    },
-    pre () {
-      return this.scaffold.pre
-    },
-    chars () {
-      return this.scaffold.chars
-    },
-    validationMessage () {
-      return this.scaffold.validationMessage
-    },
-    value () {
-      return this.field.value
-    },
-    originalValue () {
-      return this.field.originalValue
-    },
-    min () {
-      return this.scaffold.min
-    },
-    max () {
-      return this.scaffold.max
-    },
-    empty () {
-      const value = this.value
-      return !value || value === ''
-    },
-    state () {
-      return this.field.state
-    },
-    predictive () {
-      return this.scaffold.predictive
-    },
-    predictiveOptions () {
-      return this.scaffold.predictiveOptions
-    },
-    predictiveDisplayKey () {
-      return this.scaffold.predictiveOptions.display_key
-    },
-    predictiveOptionsList () {
-      return this.scaffold.predictiveOptions.options
-    },
-    predictiveFiltered () {
-      const filterValue = this.value.toLowerCase()
-      const key = this.predictiveDisplayKey
-      const filtered = this.predictiveOptionsList.filter((obj) => {
-        const value = obj[key] ? obj[key].toLowerCase() : ''
-        if (value.includes(filterValue)) { return obj }
-        return false
-      })
-      if (filtered.length === 0) { return false }
-      return filtered
-    }
-  },
-
-  watch: {
-    value (value) {
-      preValidate(this, value, this.pre)
-    }
-  },
-
-  methods: {
-    toggleFocused (focused) {
-      this.focused = focused
-      if (focused) {
-        this.openDropdown()
-      }
-      this.$emit('toggleFocused', focused)
-    },
-    openDropdown () {
-      this.dropdownOpen = true
-    },
-    closeDropdown () {
-      this.dropdownOpen = false
-    },
-    selected (option) {
-      return option[this.predictiveDisplayKey] === this.value
-    },
-    selectOption (option) {
-      this.$emit('updateValue', option[this.predictiveDisplayKey])
-      this.closeDropdown()
-    }
   }
 }
 </script>
@@ -194,26 +37,29 @@ $height: 4rem;
 // ///////////////////////////////////////////////////////////////////// General
 .field-input {
   height: $height;
-  &.focused.predictive.dropdown-open {
-    .input {
-      border-bottom-color: transparent;
+  &.caution {
+    :deep(.input) {
+      border-color: $mandysPink;
+    }
+  }
+  &.error {
+    :deep(.input) {
+      border-color: $flamingo;
+    }
+  }
+  &.disabled {
+    cursor: no-drop;
+    :deep(.input) {
+      border-bottom-color: rgba(246, 245, 255, 0.25);
     }
   }
 }
 
-.input-container,
-.input {
-  height: 100%;
-}
-
-.input {
+:deep(.input) {
   display: flex;
   flex-direction: row;
   align-items: center;
-  width: 100%;
   border-bottom: 2px solid $titanWhite;
-  appearance: none;
-  -moz-appearance: textfield;
   transition: 150ms ease-in-out;
   @include placeholder {
     color: rgba($aquaSqueeze, 0.7);
@@ -230,74 +76,5 @@ $height: 4rem;
       font-size: toRem(20);
     }
   }
-  &::-webkit-inner-spin-button,
-  &::-webkit-outer-spin-button {
-    margin: 0;
-    appearance: none;
-  }
-  &.caution {
-    border-color: $mandysPink;
-  }
-  &.error {
-    border-color: $flamingo;
-  }
-  &.disabled {
-    cursor: no-drop;
-    border-bottom-color: rgba(246, 245, 255, 0.25);
-  }
-}
-
-// ///////////////////////////////////////////////////////// Predictive Dropdown
-.dropdown {
-  position: absolute;
-  top: calc(100% - 2px);
-  left: 0;
-  width: 100%;
-  max-height: $height * 5.5;
-  background-color: $aztec;
-  border: 2px solid $titanWhite;
-  border-radius: 0.3125rem;
-  opacity: 0;
-  visibility: hidden;
-  pointer-events: none;
-  overflow: hidden;
-  z-index: -1;
-  &.open {
-    opacity: 1;
-    visibility: visible;
-    pointer-events: all;
-    overflow-y: scroll;
-    z-index: 100;
-  }
-}
-
-.options-container {
-  overflow: hidden;
-  width: 100%;
-  height: 100%;
-  z-index: 5;
-}
-
-.option {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  width: 100%;
-  height: $height;
-  padding: 0 0.875rem;
-  color: inherit;
-  &:hover,
-  &.selected {
-    background-color: rgba(white, 0.1);
-  }
-}
-
-.none-found-placeholder {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: $height * 2;
-  padding: 2rem;
 }
 </style>
